backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.5.2
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==24.1.0
face-recognition==1.3.0
numpy==1.26.4
pillow==10.4.0
